Fix loading data json
Update message format

+++ src/main/kotlin/com/subakstudio/sandbox/cafenotifier/search/CafeArticle.kt
@@ -3,6 +3,6 @@ package com.subakstudio.sandbox.cafenotifier.search
 /**
  * Created by yeoupooh on 16. 4. 19.
  */
-data class CafeArticle(var id: Int, var title: String, var name: String, var time: String, var views: Int, var href: String, var text: String) {
+data class CafeArticle(var id: Long, var title: String, var name: String, var time: String, var views: Int, var href: String, var text: String) {
 
 }
